Improve favorites UI
