+++ src/main/java/seedu/mark/ui/DashboardPanel.java
@@ -3,18 +3,13 @@ package seedu.mark.ui;
 import java.util.function.Consumer;
 import java.util.logging.Logger;
 
-import javafx.collections.ObservableList;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.StackPane;
 import seedu.mark.commons.core.LogsCenter;
 import seedu.mark.logic.Logic;
 import seedu.mark.model.bookmark.Url;
-import javafx.scene.input.MouseEvent;
-import seedu.mark.model.bookmark.Bookmark;
+
 
 /**
  * The Dashboard panel of Mark.
@@ -27,10 +22,10 @@ public class DashboardPanel extends UiPart<Region> {
     @FXML
     private StackPane folderStructurePlaceholder;
     @FXML
-    private StackPane reminderListPlaceholder;
+    private StackPane favoriteListPlaceholder;
     @FXML
-    private FlowPane favoriteBookmarkList;
-    private ObservableList<Bookmark> favoriteBookmarks;
+    private StackPane reminderListPlaceholder;
+
 
     public DashboardPanel(Logic logic, Consumer<Url> currentUrlChangeHandler) {
         super(FXML);
@@ -38,18 +33,12 @@ public class DashboardPanel extends UiPart<Region> {
                 logic.getFolderStructure(), logic.getFilteredBookmarkList(), currentUrlChangeHandler);
         this.folderStructureTreeView = folderStructureTreeView;
         folderStructurePlaceholder.getChildren().add(folderStructureTreeView.getRoot());
+
         ReminderListPanel reminderListPanel = new ReminderListPanel(logic.getReminderList());
         reminderListPlaceholder.getChildren().add(reminderListPanel.getRoot());
-        favoriteBookmarks = logic.getFavoriteBookmarkList();
-        favoriteBookmarks.forEach(bookmark -> {
-            Label bookmarkLabel = new Label(bookmark.getName().value);
-            bookmarkLabel.setOnMouseClicked(new EventHandler<MouseEvent>() {
-                @Override
-                public void handle(MouseEvent arg0) {
-                    currentUrlChangeHandler.accept(bookmark.getUrl());
-                }
-            });
-            favoriteBookmarkList.getChildren().add(bookmarkLabel);
-        });
+
+        FavoriteListPanel favoriteListPanel = new FavoriteListPanel(logic.getFavoriteBookmarkList(),
+                currentUrlChangeHandler);
+        favoriteListPlaceholder.getChildren().add(favoriteListPanel.getRoot());
     }
 }
