+++ src/main/resources/view/FolderStructure.fxml
@@ -4,7 +4,6 @@
 
 <TreeView xmlns="http://javafx.com/javafx/8"
   xmlns:fx="http://javafx.com/fxml"
-  prefHeight="400.0" prefWidth="600.0"
   fx:id="treeView">
 
 </TreeView>
